rename test files for consistency
