+++ lib/index.js
@@ -34,9 +34,7 @@ internals.implementation = function (server, options) {
       if (!extract.isValid(token)) { // quick check for validity of token format
         return reply(Boom.unauthorized('Invalid token format', 'Token'));
       } // verification is done later, but we want to avoid decoding if malformed
-
-      // if the keyFunc is a function it allows dynamic key lookup see: https://github.com/dwyl/hapi-auth-jwt2/issues/16
-      var keyFunc = (internals.isFunction(options.key)) ? options.key : function (decoded, callback) { callback(null, options.key); };
+      request.auth.token = token; // keep encoded JWT available in the request lifecycle
       // otherwise use the same key (String) to validat all JWTs
       var decoded;
       try {
@@ -50,40 +48,56 @@ internals.implementation = function (server, options) {
         return reply.continue({ credentials: decoded, artifacts: token });
       } // make token available in request.auth.credentials but skip JWT.verify
 
-      keyFunc(decoded, function (err, key, extraInfo) {
-        if (err) {
-          return reply(Boom.wrap(err));
-        }
-        if (extraInfo) {
-          request.plugins[pkg.name] = { extraInfo: extraInfo };
-        }
-        // additional checks
-        assert(options.key, 'options must contain secret key or key lookup function'); // no signing key
-        assert(typeof options.validateFunc === 'function', 'options.validateFunc must be a valid function');
-        var verifyOptions = options.verifyOptions || {};
-        JWT.verify(token, key, verifyOptions, function (err, decoded) {
-          if (err && err.name === 'TokenExpiredError') {
-            return reply(Boom.unauthorized('Token expired', 'Token'), null, { credentials: null });
+      if(options.key && typeof options.validateFunc === 'function') {
+        // if the keyFunc is a function it allows dynamic key lookup see: https://github.com/dwyl/hapi-auth-jwt2/issues/16
+        var keyFunc = (internals.isFunction(options.key)) ?
+        options.key : function (decoded, callback) { callback(null, options.key); };
+
+        keyFunc(decoded, function (err, key, extraInfo) {
+          if (err) {
+            return reply(Boom.wrap(err));
           }
-          else if (err) {
-            return reply(Boom.unauthorized('Invalid token', 'Token'), null, { credentials: null });
+          if (extraInfo) {
+            request.plugins[pkg.name] = { extraInfo: extraInfo };
           }
-          else { // see: http://hapijs.com/tutorials/auth for validateFunc signature
-            options.validateFunc(decoded, request, function (err, valid, credentials) { // bring your own checks
-              if (err) {
-                return reply(Boom.wrap(err));
-              }
-              else if (!valid) {
-                return reply(Boom.unauthorized('Invalid credentials', 'Token'), null, { credentials: credentials || decoded });
-              }
-              else {
-                request.auth.token = token;
-                return reply.continue({ credentials: credentials || decoded, artifacts: token });
-              }
-            });
+          var verifyOptions = options.verifyOptions || {};
+          JWT.verify(token, key, verifyOptions, function (err, decoded) {
+            if (err && err.name === 'TokenExpiredError') {
+              return reply(Boom.unauthorized('Token expired', 'Token'), null, { credentials: null });
+            }
+            else if (err) {
+              return reply(Boom.unauthorized('Invalid token', 'Token'), null, { credentials: null });
+            }
+            else { // see: http://hapijs.com/tutorials/auth for validateFunc signature
+              options.validateFunc(decoded, request, function (err, valid, credentials) { // bring your own checks
+                if (err) {
+                  return reply(Boom.wrap(err));
+                }
+                else if (!valid) {
+                  return reply(Boom.unauthorized('Invalid credentials', 'Token'), null, { credentials: credentials || decoded });
+                }
+                else {
+                  return reply.continue({ credentials: credentials || decoded, artifacts: token });
+                }
+              });
+            }
+          });
+        }); // END keyFunc
+      } // END check for (secret) key and validateFunc
+      else if (typeof options.customVerify === 'function') { // see: https://github.com/dwyl/hapi-auth-jwt2/issues/130
+        options.customVerify(decoded, request, function(err, valid, credentials) {
+          if (err) {
+            return reply(Boom.wrap(err));
+          }
+          else if (!valid) {
+            return reply(Boom.unauthorized('Invalid credentials', 'Token'), null, { credentials: credentials || decoded });
+          } else {
+            return reply.continue({ credentials: credentials, artifacts: token });
           }
         });
-      });
+      } else {
+        return reply.continue({ credentials: decoded, artifacts: token });
+      } // make token available in request.auth.credentials but skip JWT.verify
     }
   };
 };
